Improve displaying conditions information on details view
