+++ core-ui/src/components/Predefined/List/EventSubscriptions.list.js
@@ -13,12 +13,7 @@ export const SubscriptionsList = ({ DefaultRenderer, ...otherParams }) => {
       value: ({ status }) => {
         const lastCondition = status.conditions[status.conditions.length - 1];
 
-        return (
-          <EventSubscriptionConditionStatus
-            condition={lastCondition}
-            isListView={true}
-          />
-        );
+        return <EventSubscriptionConditionStatus condition={lastCondition} />;
       },
     },
   ];
