+++ core/src/i18n/en.yaml
@@ -782,7 +782,7 @@ domains:
     placeholder: Domain that is allowed
 event-subscription:
   conditions:
-    last-transition-time: Last transition time
+    last-transition-time: Last transition
     reason: Reason
     status: Status
     title: Conditions
